requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
gql==3.0.0a2
graphql-core==3.1.2
idna==2.10
importlib-metadata==2.0.0
jsonpickle==1.4.1
multidict==4.7.6
polog==0.0.6
pony==0.7.13
promise==2.3
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
Rx==1.6.1
six==1.15.0
tornado==6.0.4
tzlocal==2.1
ujson==3.0.0
urllib3==1.25.10
websockets==8.1
yarl==1.6.2
zipp==3.3.1
